restful_snmp_api/requirements.txt
flask==2.1.0
PyYaml==5.4.1
apscheduler==3.7.0
easysnmp==0.2.5
